Documentation: coreboot Gerrit Etiquette and Guidelines
As the community has grown, so has the need to formalize some of the
guidelines that the community lives by.  When the community was small,
it was easy to communicate these things just from one person to another.

Now, with more people joining the community every day, it seems that
it's time to write some of these things down, allowing people to
understand our policies immediately instead of making them learn our
practices as they make mistakes.

As it says in the document:
The following rules are the requirements for behavior in the coreboot
codebase in gerrit.  These have mainly been unwritten rules up to this
point, and should be familiar to most users who have been active in
coreboot for a period of time.  Following these rules will help reduce
friction in the community.

Change-Id: If80e933fcfb04b86fd5efe6423cda448118d7a3c
Signed-off-by: Martin Roth <martinroth@google.com>
Reviewed-on: http://review.coreboot.org/12256
Tested-by: build bot (Jenkins)
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>
Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-by: Ronald G. Minnich <rminnich@gmail.com>